added base_uploader and had other uploaders inherit it
